User, Video model 연관관계 구현

+++ src/contorllers/videoController.js
@@ -1,4 +1,5 @@
 import { render } from "pug";
+import User from "../models/User";
 import Video from "../models/Video";
 
 // export const home = async (req, res) => {
@@ -14,7 +15,8 @@ export const home = async(req, res) => {
 
 export const watch = async(req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    // populate("owner") : owner로 포함한 User 정보를 모두 가져온다.
+    const video = await Video.findById(id).populate("owner");
     if (!video) {
         return res.render("404", { pageTitle: "Video not found." });
     }
@@ -24,19 +26,33 @@ export const watch = async(req, res) => {
 export const getEdit = async(req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id);
+    const {
+        user: {_id}
+    } = req.session;
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
 };
 
 export const postEdit = async(req, res) => {
+    const {
+        user: {_id}
+    } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
+
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndUpdate(id, {
         title,
         description,
@@ -50,13 +66,24 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async(req, res) => {
+    const {
+        user: {_id},
+    } = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
+    
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags)
         });
+        // video를 생성할 때 user의 video object에도 넣어준다.
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
     } catch (error) {
         console.log(error);
         return res.status(400).render("upload", {
@@ -64,11 +91,23 @@ export const postUpload = async(req, res) => {
             errorMessage: `😓 ${error._message}`
         });
     };
-    // const dbVideo = await video.save();
     return res.redirect("/");
 };
 
 export const deleteVideo = async(req, res) => {
+    const {
+        user: {_id}
+    } = req.session;
+
+    const video = await Video.findById(_id);
+    if(!video){
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
+
     const { id } = req.params;
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
